requirements/dev.txt
################################################
## DO NOT EDIT - AUTOMATICALLY GENERATED FILE ##
################################################
astroid==1.4.5
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
cffi==1.6.0
click==3.3
cloverpy==0.1.1
cov-core==1.15.0
coverage==4.0.3
cryptography==1.1
docopt==0.6.2
docutils==0.12
enum34==1.1.4
git+https://github.com/catapultbamboo/file_translate.git@0.0.3
Flask==0.10.1
funcsigs==1.0.2
futures==3.0.3
grip==3.2.0
html2text==3.2
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
lazy-object-proxy==1.2.2
logilab-common==1.2.0
Markdown==2.6.6
MarkupSafe==0.23
mock==1.3.0
oauth==1.0.1
opp-chub==1.0.5
opp-koi==1.0.8
git+https://github.com/openpermissions/perch.git@0.14.0
passlib==1.6.2
path-and-address==1.1.0
pbr==1.9.1
py==1.4.31
pyasn1==0.1.9
pycparser==2.14
Pygments==2.1.3
pylint==1.3.0
pytest==2.6.3
pytest-cov==1.8.0
python-dateutil==2.5.2
requests==2.4.3
singledispatch==3.4.0.3
six==1.10.0
Sphinx==1.2.3
tornado==4.3
tornado-couchdb==0.2.3
voluptuous==0.8.9
Werkzeug==0.11.9
wheel==0.24.0
wrapt==1.10.8
